backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.7.0
sentence-transformers>=2.2.2
chromadb>=0.4.15
langchain>=0.1.0
langchain-community>=0.0.10
google-generativeai>=0.3.0
pypdf2>=3.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1

